Refactor action card layout helpers
